apps/api/pyproject.toml
[project]
name = "infralens-api"
version = "0.1.0"
description = "FastAPI backend for InfraLens Georgia."
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "pgvector>=0.3.0",
    "psycopg[binary]>=3.2.0",
    "pydantic-settings>=2.5.0",
    "redis>=5.0.0",
    "rq>=2.0.0",
    "sqlalchemy>=2.0.0",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
]

[build-system]
requires = ["setuptools>=72.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

